requirements.txt
pyyaml==5.4.1
requests==2.28.1
